requirements/dev-requirements.txt
# -*- coding: utf-8 -*-

-r common-requirements.txt
coverage==3.6
django-nose==1.2
django-snippetscream==0.0.7
ipdb==0.7
ipython==0.13.2
Jinja2==2.7
logilab-astng==0.24.3
logilab-common==0.59.1
nose==1.3.0
pip-tools==0.3.4
Pygments==1.6
# Activate the following if you need fancy graph drawing for the dependencies
pygraphviz==1.1
pylint==0.27.0
pyflakes>=0.7.3
Sphinx>=1.2b1
Werkzeug==0.9.3
